🗺️ Atlas: Splitting Property Blob module
