+++ src/core/property/builder.rs
@@ -1,0 +1,223 @@
+use std::collections::HashMap;
+use std::hash::BuildHasherDefault;
+use std::sync::Arc;
+
+use crate::core::error::Result;
+use crate::core::hasher::IdentityHasher;
+use crate::core::interning::GLOBAL_INTERNER;
+
+use super::*;
+
+/// Builder for creating or modifying property maps with copy-on-write semantics.
+pub struct PropertyMapBuilder {
+    pub(crate) map: HashMap<PropertyKey, PropertyValue, BuildHasherDefault<IdentityHasher>>,
+    current_size: usize,
+}
+
+impl PropertyMapBuilder {
+    /// Create a new builder with an empty map.
+    pub fn new() -> Self {
+        PropertyMapBuilder {
+            map: HashMap::with_hasher(BuildHasherDefault::default()),
+            current_size: 4, // Count field
+        }
+    }
+
+    /// Create a builder from an existing PropertyMap.
+    ///
+    /// This will clone the underlying HashMap if the Arc has multiple references,
+    /// implementing copy-on-write semantics.
+    pub fn from_map(prop_map: PropertyMap) -> Self {
+        let current_size = prop_map.cached_size;
+        let map = Arc::try_unwrap(prop_map.inner).unwrap_or_else(|arc: Arc<HashMap<PropertyKey, PropertyValue, BuildHasherDefault<IdentityHasher>>>| (*arc).clone());
+        PropertyMapBuilder { map, current_size }
+    }
+
+    /// Insert a property.
+    ///
+    /// The key is automatically interned. If interning fails (capacity exceeded),
+    /// returns self unchanged.
+    ///
+    /// Panics if recursion depth limit is exceeded.
+    pub fn insert<V: Into<PropertyValue>>(self, key: &str, value: V) -> Self {
+        self.try_insert(key, value)
+            .expect("Property insertion failed (recursion depth limit exceeded)")
+    }
+
+    /// Insert a property (fallible).
+    pub fn try_insert<V: Into<PropertyValue>>(mut self, key: &str, value: V) -> Result<Self> {
+        // Warden: Propagate interning errors (e.g. CapacityExceeded) to prevent silent data loss.
+        // Previously, failure to intern would silently drop the property, which is a security risk.
+        let interned_key = GLOBAL_INTERNER.intern(key)?;
+        let val = value.into();
+        let val_size = val.serialized_size()?;
+
+        if let Some(old_val) = self.map.insert(interned_key, val) {
+            // Replaced existing entry
+            // Key size is unchanged (same key ID means same string)
+            self.current_size = self
+                .current_size
+                .saturating_sub(old_val.serialized_size()?)
+                .saturating_add(val_size);
+        } else {
+            // New entry
+            let key_size = 4 + key.len(); // Length prefix (4) + string bytes
+            self.current_size = self
+                .current_size
+                .saturating_add(key_size)
+                .saturating_add(val_size);
+        }
+        Ok(self)
+    }
+
+    /// Insert a property with an already-interned key.
+    ///
+    /// Panics if recursion depth limit is exceeded.
+    pub fn insert_by_key(self, key: PropertyKey, value: PropertyValue) -> Self {
+        self.try_insert_by_key(key, value)
+            .expect("Property insertion failed (recursion depth limit exceeded)")
+    }
+
+    /// Insert a property with an already-interned key (fallible).
+    pub fn try_insert_by_key(mut self, key: PropertyKey, value: PropertyValue) -> Result<Self> {
+        let val_size = value.serialized_size()?;
+
+        if let Some(old_val) = self.map.insert(key, value) {
+            // Replaced existing entry - key size constant
+            self.current_size = self
+                .current_size
+                .saturating_sub(old_val.serialized_size()?)
+                .saturating_add(val_size);
+        } else {
+            // New entry - need key size!
+            // We must look up the string length since we only have the ID.
+            // This is a tradeoff: we pay lookup cost for new keys, but
+            // avoid it for updates and for subsequent serialization size checks.
+            let key_len = GLOBAL_INTERNER
+                .resolve_with(key, |s| s.len())
+                .unwrap_or_else(|| {
+                    // This should be unreachable if the PropertyKey is valid (which it should be).
+                    // In debug builds, we panic to catch this state corruption.
+                    // In release, we fallback to a safe estimate (256 bytes) to avoid crashing.
+                    debug_assert!(false, "PropertyKey {} missing from interner", key.as_u32());
+                    256
+                });
+            let key_size = 4 + key_len;
+            self.current_size = self
+                .current_size
+                .saturating_add(key_size)
+                .saturating_add(val_size);
+        }
+        Ok(self)
+    }
+
+    /// Insert a vector property (convenience method for embeddings).
+    ///
+    /// This is a convenience wrapper around `insert()` for vector properties,
+    /// commonly used for storing embeddings in nodes and edges.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use aletheiadb::core::property::PropertyMapBuilder;
+    ///
+    /// let embedding = vec![0.1f32, 0.2, 0.3, 0.4];
+    /// let props = PropertyMapBuilder::new()
+    ///     .insert("name", "Document")
+    ///     .insert_vector("embedding", &embedding)
+    ///     .build();
+    ///
+    /// assert_eq!(
+    ///     props.get("embedding").and_then(|v| v.as_vector()),
+    ///     Some(&embedding[..])
+    /// );
+    /// ```
+    pub fn insert_vector(self, key: &str, vector: &[f32]) -> Self {
+        self.insert(key, PropertyValue::vector(vector))
+    }
+
+    /// Insert a vector property (fallible).
+    pub fn try_insert_vector(self, key: &str, vector: &[f32]) -> Result<Self> {
+        self.try_insert(key, PropertyValue::try_vector(vector)?)
+    }
+
+    /// Remove a property.
+    ///
+    /// The key is automatically interned before removal.
+    /// If interning fails (capacity exceeded), returns self unchanged.
+    ///
+    /// Panics if serialization size calculation fails.
+    pub fn remove(self, key: &str) -> Self {
+        self.try_remove(key).expect("Property removal failed")
+    }
+
+    /// Remove a property (fallible).
+    pub fn try_remove(self, key: &str) -> Result<Self> {
+        let Some(interned_key) = GLOBAL_INTERNER.get_id(key) else {
+            return Ok(self);
+        };
+        self.try_remove_by_key(&interned_key)
+    }
+
+    /// Remove a property by an already-interned key.
+    ///
+    /// Panics if serialization size calculation fails.
+    pub fn remove_by_key(self, key: &PropertyKey) -> Self {
+        self.try_remove_by_key(key)
+            .expect("Property removal failed")
+    }
+
+    /// Remove a property by an already-interned key (fallible).
+    pub fn try_remove_by_key(mut self, key: &PropertyKey) -> Result<Self> {
+        let old_val = self.map.remove(key);
+        if let Some(old_val) = old_val {
+            let key_len = GLOBAL_INTERNER
+                .resolve_with(*key, |s| s.len())
+                .unwrap_or(256);
+            let key_size = 4 + key_len;
+            self.current_size = self
+                .current_size
+                .saturating_sub(key_size)
+                .saturating_sub(old_val.serialized_size()?);
+        }
+        Ok(self)
+    }
+
+    /// Build the final PropertyMap.
+    pub fn build(self) -> PropertyMap {
+        PropertyMap {
+            inner: Arc::new(self.map),
+            cached_size: self.current_size,
+        }
+    }
+}
+
+impl Default for PropertyMapBuilder {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+/// Macro for creating property maps with a convenient syntax.
+///
+/// # Examples
+///
+/// ```ignore
+/// let props = properties! {
+///     "name" => "Alice",
+///     "age" => 30,
+///     "active" => true,
+/// };
+/// ```
+#[macro_export]
+macro_rules! properties {
+    ($($key:expr => $value:expr),* $(,)?) => {
+        {
+            let mut builder = $crate::core::property::PropertyMapBuilder::new();
+            $(
+                builder = builder.insert($key, $value);
+            )*
+            builder.build()
+        }
+    };
+}
