+++ src/core/property/value.rs
@@ -1,0 +1,1068 @@
+use std::fmt;
+use std::sync::Arc;
+
+use crate::core::error::{Result, StorageError};
+use crate::core::vector::SparseVec;
+use crate::core::vector::serialization::validate_vector_dimensions;
+
+use super::*;
+
+/// A value that can be stored as a property.
+///
+/// All complex types (strings, bytes, arrays) use Arc for cheap cloning and sharing.
+#[derive(Clone)]
+pub enum PropertyValue {
+    /// Null/absent value.
+    Null,
+    /// Boolean value.
+    Bool(bool),
+    /// 64-bit signed integer.
+    Int(i64),
+    /// 64-bit floating point.
+    Float(f64),
+    /// UTF-8 string (reference counted).
+    String(Arc<str>),
+    /// Byte array (reference counted).
+    Bytes(Arc<[u8]>),
+    /// Array of values (reference counted).
+    Array(Arc<Vec<PropertyValue>>),
+    /// Dense vector for embeddings (reference counted).
+    /// Uses f32 for memory efficiency - standard for ML embeddings.
+    ///
+    /// # Floating-Point Equality Note
+    /// This variant uses derived PartialEq which compares f32 values bitwise.
+    /// Be aware that NaN != NaN (IEEE 754) and floating-point precision may
+    /// cause semantically equal vectors to compare unequal. For similarity
+    /// comparisons, use dedicated vector utility functions (e.g., cosine similarity)
+    /// rather than equality. This limitation will be revisited in Phase 3 when
+    /// vectors are used in temporal storage for deduplication.
+    Vector(Arc<[f32]>),
+    /// Sparse vector for high-dimensional sparse embeddings (reference counted).
+    /// Stores only non-zero values along with their indices, making it memory-efficient
+    /// for vectors where most values are zero (e.g., BM25, SPLADE).
+    ///
+    /// # Use Cases
+    /// - BM25 text retrieval vectors
+    /// - SPLADE sparse learned embeddings
+    /// - TF-IDF document vectors
+    /// - One-hot categorical encodings
+    ///
+    /// # Memory Efficiency
+    /// For a 10,000-dimensional vector with 10 non-zero values:
+    /// - Dense: 40KB (10,000 * 4 bytes)
+    /// - Sparse: ~80 bytes (10 * 8 bytes for index+value pairs)
+    /// - Space savings: ~500x
+    ///
+    /// # Floating-Point Equality Note
+    /// This variant uses derived PartialEq which compares f32 values bitwise.
+    /// Be aware that NaN != NaN (IEEE 754) and floating-point precision may
+    /// cause semantically equal vectors to compare unequal. For robust equality
+    /// checks, use [`SparseVec::approx_eq`](crate::core::vector::SparseVec::approx_eq)
+    /// with an appropriate epsilon value instead of direct `==` comparison.
+    SparseVector(Arc<SparseVec>),
+}
+
+impl PartialEq for PropertyValue {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            (PropertyValue::Null, PropertyValue::Null) => true,
+            (PropertyValue::Bool(a), PropertyValue::Bool(b)) => a == b,
+            (PropertyValue::Int(a), PropertyValue::Int(b)) => a == b,
+            (PropertyValue::Float(a), PropertyValue::Float(b)) => a == b,
+            (PropertyValue::String(a), PropertyValue::String(b)) => Arc::ptr_eq(a, b) || a == b,
+            (PropertyValue::Bytes(a), PropertyValue::Bytes(b)) => Arc::ptr_eq(a, b) || a == b,
+            (PropertyValue::Array(a), PropertyValue::Array(b)) => Arc::ptr_eq(a, b) || a == b,
+            (PropertyValue::Vector(a), PropertyValue::Vector(b)) => Arc::ptr_eq(a, b) || a == b,
+            (PropertyValue::SparseVector(a), PropertyValue::SparseVector(b)) => {
+                Arc::ptr_eq(a, b) || a == b
+            }
+            _ => false,
+        }
+    }
+}
+
+impl PropertyValue {
+    /// Create a string property value from a &str.
+    #[inline]
+    pub fn string<S: AsRef<str>>(s: S) -> Self {
+        PropertyValue::String(Arc::from(s.as_ref()))
+    }
+
+    /// Create a bytes property value from a slice.
+    #[inline]
+    pub fn bytes<B: AsRef<[u8]>>(b: B) -> Self {
+        PropertyValue::Bytes(Arc::from(b.as_ref()))
+    }
+
+    /// Create an array property value from a Vec.
+    #[inline]
+    pub fn array(values: Vec<PropertyValue>) -> Self {
+        PropertyValue::Array(Arc::new(values))
+    }
+
+    /// Create a vector property value from a slice.
+    ///
+    /// Dense vectors are used for storing embeddings in nodes and edges,
+    /// enabling semantic search and similarity computations. The data is
+    /// stored in an `Arc<[f32]>` for efficient cloning and sharing across
+    /// versions.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use aletheiadb::core::PropertyValue;
+    ///
+    /// // From a Vec
+    /// let embedding = vec![0.1f32, 0.2, 0.3, 0.4];
+    /// let prop = PropertyValue::vector(&embedding);
+    ///
+    /// // From a slice
+    /// let prop2 = PropertyValue::vector(&[0.5f32, 0.6, 0.7]);
+    ///
+    /// // Retrieve the vector
+    /// assert_eq!(prop.as_vector(), Some(&embedding[..]));
+    /// ```
+    ///
+    /// # Performance
+    ///
+    /// - Cloning a `PropertyValue::Vector` is O(1) (just increments Arc refcount)
+    /// - Unchanged vectors across versions share the same allocation
+    /// - Typical embedding sizes (384-4096 dims) use ~1.5-16KB per vector
+    ///
+    /// # See Also
+    ///
+    /// - [`PropertyMapBuilder::insert_vector`] for a builder-pattern alternative
+    /// - [`as_vector`](Self::as_vector) for retrieving the vector data
+    /// - [`aletheiadb::core::vector`](crate::core::vector) for similarity functions
+    ///
+    /// # Panics
+    ///
+    /// Panics if the vector dimension exceeds [`MAX_VECTOR_DIMENSIONS`].
+    /// This validation ensures that vectors can be serialized without error.
+    ///
+    /// For a fallible version that returns `Result` instead of panicking,
+    /// use [`try_vector`](Self::try_vector).
+    #[inline]
+    pub fn vector<V: AsRef<[f32]>>(v: V) -> Self {
+        Self::try_vector(v).unwrap_or_else(|e| panic!("{}", e))
+    }
+
+    /// Create a vector property value from a slice (fallible).
+    ///
+    /// This is the fallible version of [`vector`](Self::vector). It returns
+    /// an error if the vector dimension exceeds [`MAX_VECTOR_DIMENSIONS`]
+    /// instead of panicking.
+    #[inline]
+    pub fn try_vector<V: AsRef<[f32]>>(v: V) -> Result<Self> {
+        let slice = v.as_ref();
+        validate_vector_dimensions(slice.len())?;
+        Ok(PropertyValue::Vector(Arc::from(slice)))
+    }
+
+    /// Returns true if this value is null.
+    #[inline]
+    pub const fn is_null(&self) -> bool {
+        matches!(self, PropertyValue::Null)
+    }
+
+    /// Try to get this value as a bool.
+    #[inline]
+    pub const fn as_bool(&self) -> Option<bool> {
+        match self {
+            PropertyValue::Bool(b) => Some(*b),
+            _ => None,
+        }
+    }
+
+    /// Try to get this value as an integer.
+    #[inline]
+    pub const fn as_int(&self) -> Option<i64> {
+        match self {
+            PropertyValue::Int(i) => Some(*i),
+            _ => None,
+        }
+    }
+
+    /// Try to get this value as a float.
+    #[inline]
+    pub const fn as_float(&self) -> Option<f64> {
+        match self {
+            PropertyValue::Float(f) => Some(*f),
+            _ => None,
+        }
+    }
+
+    /// Try to get this value as a string reference.
+    #[inline]
+    pub fn as_str(&self) -> Option<&str> {
+        match self {
+            PropertyValue::String(s) => Some(s.as_ref()),
+            _ => None,
+        }
+    }
+
+    /// Try to get this value as a byte slice.
+    #[inline]
+    pub fn as_bytes(&self) -> Option<&[u8]> {
+        match self {
+            PropertyValue::Bytes(b) => Some(b.as_ref()),
+            _ => None,
+        }
+    }
+
+    /// Try to get this value as an array.
+    #[inline]
+    pub fn as_array(&self) -> Option<&[PropertyValue]> {
+        match self {
+            PropertyValue::Array(a) => Some(a.as_ref()),
+            _ => None,
+        }
+    }
+
+    /// Try to get this value as a vector (dense embedding).
+    ///
+    /// Returns `Some(&[f32])` if this is a `Vector` variant, `None` otherwise.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use aletheiadb::core::PropertyValue;
+    ///
+    /// let embedding = vec![0.1f32, 0.2, 0.3];
+    /// let prop = PropertyValue::vector(&embedding);
+    ///
+    /// if let Some(vec) = prop.as_vector() {
+    ///     assert_eq!(vec.len(), 3);
+    ///     assert!((vec[0] - 0.1).abs() < f32::EPSILON);
+    /// }
+    ///
+    /// // Returns None for non-vector types
+    /// let int_prop = PropertyValue::Int(42);
+    /// assert!(int_prop.as_vector().is_none());
+    /// ```
+    ///
+    /// # See Also
+    ///
+    /// - [`vector`](Self::vector) for creating vector properties
+    /// - [`aletheiadb::core::vector`](crate::core::vector) for similarity functions
+    #[inline]
+    pub fn as_vector(&self) -> Option<&[f32]> {
+        match self {
+            PropertyValue::Vector(v) => Some(v.as_ref()),
+            _ => None,
+        }
+    }
+
+    /// Get the underlying Arc for a vector (dense embedding) without copying.
+    ///
+    /// Returns `Some(Arc<[f32]>)` if this is a `Vector` variant, `None` otherwise.
+    /// This is more efficient than `as_vector().map(|s| s.to_vec())` because it
+    /// clones the Arc (O(1) reference count increment) rather than copying the
+    /// entire vector data.
+    ///
+    /// # Use Case
+    ///
+    /// Use this method when you need an owned reference to the vector data that
+    /// can outlive the PropertyValue, without incurring the cost of copying.
+    /// This is particularly useful for:
+    /// - Passing vectors to functions that need ownership
+    /// - Storing vector references across async boundaries
+    /// - Avoiding allocations in performance-critical paths (Issue #188)
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use aletheiadb::core::PropertyValue;
+    /// use std::sync::Arc;
+    ///
+    /// let embedding = vec![0.1f32, 0.2, 0.3];
+    /// let prop = PropertyValue::vector(&embedding);
+    ///
+    /// // Get an Arc to the data without copying
+    /// if let Some(arc) = prop.as_arc_vector() {
+    ///     assert_eq!(arc.len(), 3);
+    ///     // The Arc can outlive `prop` and be passed around cheaply
+    /// }
+    /// ```
+    ///
+    /// # Performance
+    ///
+    /// - O(1) operation (just increments Arc reference count)
+    /// - No memory allocation or data copying
+    /// - Safe to call multiple times (returns same underlying data)
+    ///
+    /// # See Also
+    ///
+    /// - [`as_vector`](Self::as_vector) for borrowing the vector as a slice
+    /// - [`vector`](Self::vector) for creating vector properties
+    #[inline]
+    pub fn as_arc_vector(&self) -> Option<Arc<[f32]>> {
+        match self {
+            PropertyValue::Vector(v) => Some(Arc::clone(v)),
+            _ => None,
+        }
+    }
+
+    /// Create a sparse vector property value from a SparseVec.
+    ///
+    /// Sparse vectors store only non-zero values along with their indices,
+    /// making them memory-efficient for high-dimensional vectors where most
+    /// values are zero (e.g., BM25, SPLADE).
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use aletheiadb::core::PropertyValue;
+    /// use aletheiadb::core::vector::SparseVec;
+    ///
+    /// // Create a sparse vector: [0.0, 1.5, 0.0, 0.0, 2.3, 0.0]
+    /// let sparse = SparseVec::new(vec![1, 4], vec![1.5, 2.3], 6).unwrap();
+    /// let prop = PropertyValue::sparse_vector(sparse);
+    ///
+    /// // Retrieve the sparse vector
+    /// assert!(prop.as_sparse_vector().is_some());
+    /// ```
+    ///
+    /// # Performance
+    ///
+    /// - Cloning a `PropertyValue::SparseVector` is O(1) (just increments Arc refcount)
+    /// - Memory usage: O(nnz) where nnz = number of non-zero elements
+    /// - Can save 10-1000x memory compared to dense vectors for sparse data
+    ///
+    /// # See Also
+    ///
+    /// - [`as_sparse_vector`](Self::as_sparse_vector) for retrieving the sparse vector
+    /// - [`SparseVec`] for creating sparse vectors from indices and values
+    #[inline]
+    pub fn sparse_vector(sparse: SparseVec) -> Self {
+        PropertyValue::SparseVector(Arc::new(sparse))
+    }
+
+    /// Try to get this value as a sparse vector.
+    ///
+    /// Returns `Some(&SparseVec)` if this is a `SparseVector` variant, `None` otherwise.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use aletheiadb::core::PropertyValue;
+    /// use aletheiadb::core::vector::SparseVec;
+    ///
+    /// let sparse = SparseVec::new(vec![0, 2], vec![1.0, 2.0], 5).unwrap();
+    /// let prop = PropertyValue::sparse_vector(sparse);
+    ///
+    /// if let Some(sv) = prop.as_sparse_vector() {
+    ///     assert_eq!(sv.nnz(), 2);
+    ///     assert_eq!(sv.dimension(), 5);
+    /// }
+    ///
+    /// // Returns None for non-sparse-vector types
+    /// let int_prop = PropertyValue::Int(42);
+    /// assert!(int_prop.as_sparse_vector().is_none());
+    /// ```
+    ///
+    /// # See Also
+    ///
+    /// - [`sparse_vector`](Self::sparse_vector) for creating sparse vector properties
+    /// - [`SparseVec`] for sparse vector operations
+    #[inline]
+    pub fn as_sparse_vector(&self) -> Option<&SparseVec> {
+        match self {
+            PropertyValue::SparseVector(sv) => Some(sv.as_ref()),
+            _ => None,
+        }
+    }
+
+    /// Get the type name of this value.
+    pub const fn type_name(&self) -> &'static str {
+        match self {
+            PropertyValue::Null => "null",
+            PropertyValue::Bool(_) => "bool",
+            PropertyValue::Int(_) => "int",
+            PropertyValue::Float(_) => "float",
+            PropertyValue::String(_) => "string",
+            PropertyValue::Bytes(_) => "bytes",
+            PropertyValue::Array(_) => "array",
+            PropertyValue::Vector(_) => "vector",
+            PropertyValue::SparseVector(_) => "sparse_vector",
+        }
+    }
+
+    /// Check if two property values are semantically equal.
+    ///
+    /// This differs from `PartialEq` in that it treats `NaN` values as equal.
+    /// This is important for change detection systems (like `VersionDiff` and `PropertyDelta`)
+    /// to avoid reporting spurious changes when a value remains `NaN`.
+    ///
+    /// # Handling of NaN
+    /// - `Float(NaN)` is equal to `Float(NaN)`
+    /// - `Vector` containing `NaN` at index `i` is equal to `Vector` containing `NaN` at index `i`
+    /// - `SparseVector`: Guaranteed not to contain `NaN` (enforced at construction), so standard equality applies.
+    pub fn semantically_equal(&self, other: &Self) -> bool {
+        match (self, other) {
+            (PropertyValue::Float(a), PropertyValue::Float(b)) => {
+                if a.is_nan() {
+                    b.is_nan()
+                } else {
+                    a == b
+                }
+            }
+            (PropertyValue::Vector(a), PropertyValue::Vector(b)) => {
+                // Optimization: If they point to the same allocation, they must be equal.
+                if Arc::ptr_eq(a, b) {
+                    return true;
+                }
+                if a.len() != b.len() {
+                    return false;
+                }
+                a.iter()
+                    .zip(b.iter())
+                    .all(|(x, y)| if x.is_nan() { y.is_nan() } else { x == y })
+            }
+            (PropertyValue::Array(a), PropertyValue::Array(b)) => {
+                // Optimization: If they point to the same allocation, they must be equal.
+                if Arc::ptr_eq(a, b) {
+                    return true;
+                }
+                if a.len() != b.len() {
+                    return false;
+                }
+                a.iter().zip(b.iter()).all(|(x, y)| x.semantically_equal(y))
+            }
+            // For other types, fallback to PartialEq
+            _ => self == other,
+        }
+    }
+
+    // ========================================================================
+    // Serialization Methods
+    // ========================================================================
+
+    /// Serialize this PropertyValue to bytes.
+    ///
+    /// # Binary Format
+    /// - Tag (1 byte): Identifies the value type
+    /// - Payload: Type-specific data in little-endian format
+    ///
+    /// | Type   | Format                                      |
+    /// |--------|---------------------------------------------|
+    /// | Null   | `[tag:1]`                                   |
+    /// | Bool   | `[tag:1][value:1]`                          |
+    /// | Int    | `[tag:1][i64:8]`                            |
+    /// | Float  | `[tag:1][f64:8]`                            |
+    /// | String | `[tag:1][len:4][utf8_bytes:len]`            |
+    /// | Bytes  | `[tag:1][len:4][bytes:len]`                 |
+    /// | Array  | `[tag:1][count:4][elements...]`             |
+    /// | Vector | `[tag:1][dim:4][f32_values:dim*4]`          |
+    ///
+    /// # Errors
+    /// Returns `StorageError::CorruptedData` if recursion depth exceeds limits.
+    pub fn serialize(&self) -> Result<Vec<u8>> {
+        let mut buffer = Vec::with_capacity(self.serialized_size().map_err(|_| {
+            StorageError::CorruptedData(
+                "Recursion depth limit exceeded in serialized_size".to_string(),
+            )
+        })?);
+        self.serialize_into(&mut buffer)?;
+        Ok(buffer)
+    }
+
+    /// Serialize this PropertyValue into an existing buffer.
+    ///
+    /// This is more efficient when serializing multiple values as it avoids
+    /// allocating a new Vec for each value.
+    ///
+    /// # Errors
+    /// Returns `StorageError::CorruptedData` if recursion depth exceeds limits.
+    pub fn serialize_into(&self, buffer: &mut Vec<u8>) -> Result<()> {
+        self.serialize_recursive(buffer, 0)
+    }
+
+    fn serialize_recursive(&self, buffer: &mut Vec<u8>, depth: usize) -> Result<()> {
+        if depth > MAX_RECURSION_DEPTH {
+            return Err(StorageError::CorruptedData(format!(
+                "Property value recursion depth limit exceeded (max {})",
+                MAX_RECURSION_DEPTH
+            ))
+            .into());
+        }
+
+        match self {
+            PropertyValue::Null => {
+                buffer.push(TAG_NULL);
+                Ok(())
+            }
+            PropertyValue::Bool(b) => {
+                buffer.push(TAG_BOOL);
+                buffer.push(if *b { 1 } else { 0 });
+                Ok(())
+            }
+            PropertyValue::Int(i) => {
+                buffer.push(TAG_INT);
+                buffer.extend_from_slice(&i.to_le_bytes());
+                Ok(())
+            }
+            PropertyValue::Float(f) => {
+                buffer.push(TAG_FLOAT);
+                buffer.extend_from_slice(&f.to_le_bytes());
+                Ok(())
+            }
+            PropertyValue::String(s) => {
+                buffer.push(TAG_STRING);
+                let bytes = s.as_bytes();
+                buffer.extend_from_slice(&(bytes.len() as u32).to_le_bytes());
+                buffer.extend_from_slice(bytes);
+                Ok(())
+            }
+            PropertyValue::Bytes(b) => {
+                buffer.push(TAG_BYTES);
+                buffer.extend_from_slice(&(b.len() as u32).to_le_bytes());
+                buffer.extend_from_slice(b);
+                Ok(())
+            }
+            PropertyValue::Array(arr) => {
+                buffer.push(TAG_ARRAY);
+                buffer.extend_from_slice(&(arr.len() as u32).to_le_bytes());
+                for item in arr.iter() {
+                    item.serialize_recursive(buffer, depth + 1)?;
+                }
+                Ok(())
+            }
+            PropertyValue::Vector(v) => {
+                try_serialize_vector_into(v, buffer)?;
+                Ok(())
+            }
+            PropertyValue::SparseVector(sv) => {
+                serialize_sparse_vector_into(sv, buffer);
+                Ok(())
+            }
+        }
+    }
+
+    /// Deserialize a PropertyValue from bytes.
+    ///
+    /// Returns the deserialized value and the number of bytes consumed.
+    ///
+    /// # Recursion Depth
+    ///
+    /// This function implements recursion depth checking to prevent stack overflow
+    /// attacks via deeply nested structures (e.g., Array of Array of ...).
+    /// The maximum depth is defined by [`MAX_RECURSION_DEPTH`].
+    pub fn deserialize(bytes: &[u8]) -> Result<(Self, usize)> {
+        Self::deserialize_recursive(bytes, 0)
+    }
+
+    /// Internal recursive deserialization helper with depth tracking.
+    fn deserialize_recursive(bytes: &[u8], depth: usize) -> Result<(Self, usize)> {
+        // Prevent recursion-based stack overflow DoS
+        // Depth 0 = top level, depth 100 = maximum nesting level
+        if depth > MAX_RECURSION_DEPTH {
+            return Err(StorageError::CorruptedData(format!(
+                "Property value recursion depth limit exceeded (max {})",
+                MAX_RECURSION_DEPTH
+            ))
+            .into());
+        }
+
+        if bytes.is_empty() {
+            return Err(StorageError::CorruptedData(
+                "Empty buffer when deserializing PropertyValue".to_string(),
+            )
+            .into());
+        }
+
+        let tag = bytes[0];
+
+        match tag {
+            TAG_NULL => Ok((PropertyValue::Null, 1)),
+            TAG_BOOL => Self::deserialize_bool(bytes),
+            TAG_INT => Self::deserialize_int(bytes),
+            TAG_FLOAT => Self::deserialize_float(bytes),
+            TAG_STRING => Self::deserialize_string(bytes),
+            TAG_BYTES => Self::deserialize_bytes(bytes),
+            TAG_ARRAY => Self::deserialize_array(bytes, depth),
+            TAG_VECTOR => {
+                let (vector, consumed) = deserialize_vector(bytes)?;
+                Ok((PropertyValue::Vector(vector), consumed))
+            }
+            TAG_SPARSE_VECTOR => {
+                let (sparse_vector, consumed) = deserialize_sparse_vector(bytes)?;
+                Ok((PropertyValue::SparseVector(sparse_vector), consumed))
+            }
+            _ => Err(StorageError::CorruptedData(format!(
+                "Unknown PropertyValue type tag: {}",
+                tag
+            ))
+            .into()),
+        }
+    }
+
+    fn deserialize_bool(bytes: &[u8]) -> Result<(Self, usize)> {
+        if bytes.len() < 2 {
+            return Err(
+                StorageError::CorruptedData("Buffer too short for Bool value".to_string()).into(),
+            );
+        }
+        let value = bytes[1] != 0;
+        Ok((PropertyValue::Bool(value), 2))
+    }
+
+    fn deserialize_int(bytes: &[u8]) -> Result<(Self, usize)> {
+        if bytes.len() < 9 {
+            return Err(
+                StorageError::CorruptedData("Buffer too short for Int value".to_string()).into(),
+            );
+        }
+        // SAFETY: Length check above guarantees slice has 8 bytes
+        let value = i64::from_le_bytes(bytes[1..9].try_into().unwrap());
+        Ok((PropertyValue::Int(value), 9))
+    }
+
+    fn deserialize_float(bytes: &[u8]) -> Result<(Self, usize)> {
+        if bytes.len() < 9 {
+            return Err(StorageError::CorruptedData(
+                "Buffer too short for Float value".to_string(),
+            )
+            .into());
+        }
+        // SAFETY: Length check above guarantees slice has 8 bytes
+        let value = f64::from_le_bytes(bytes[1..9].try_into().unwrap());
+        Ok((PropertyValue::Float(value), 9))
+    }
+
+    #[inline]
+    fn deserialize_string(bytes: &[u8]) -> Result<(Self, usize)> {
+        if bytes.len() < 5 {
+            return Err(StorageError::CorruptedData(
+                "Buffer too short for String length".to_string(),
+            )
+            .into());
+        }
+        let len = u32::from_le_bytes(bytes[1..5].try_into().unwrap()) as usize;
+        let offset = 5usize;
+
+        let required_len = offset
+            .checked_add(len)
+            .ok_or_else(|| StorageError::CorruptedData("String length overflow".to_string()))?;
+
+        if bytes.len() < required_len {
+            return Err(StorageError::CorruptedData(format!(
+                "Buffer too short for String data: need {} bytes, have {}",
+                required_len,
+                bytes.len()
+            ))
+            .into());
+        }
+
+        let string_data = &bytes[offset..required_len];
+        let s = std::str::from_utf8(string_data)
+            .map_err(|e| StorageError::CorruptedData(format!("Invalid UTF-8 in String: {}", e)))?;
+        Ok((PropertyValue::String(Arc::from(s)), required_len))
+    }
+
+    #[inline]
+    fn deserialize_bytes(bytes: &[u8]) -> Result<(Self, usize)> {
+        if bytes.len() < 5 {
+            return Err(StorageError::CorruptedData(
+                "Buffer too short for Bytes length".to_string(),
+            )
+            .into());
+        }
+        let len = u32::from_le_bytes(bytes[1..5].try_into().unwrap()) as usize;
+        let offset = 5usize;
+
+        let required_len = offset
+            .checked_add(len)
+            .ok_or_else(|| StorageError::CorruptedData("Bytes length overflow".to_string()))?;
+
+        if bytes.len() < required_len {
+            return Err(StorageError::CorruptedData(format!(
+                "Buffer too short for Bytes data: need {} bytes, have {}",
+                required_len,
+                bytes.len()
+            ))
+            .into());
+        }
+
+        let byte_data = &bytes[offset..required_len];
+        Ok((PropertyValue::Bytes(Arc::from(byte_data)), required_len))
+    }
+
+    fn deserialize_array(bytes: &[u8], depth: usize) -> Result<(Self, usize)> {
+        if bytes.len() < 5 {
+            return Err(StorageError::CorruptedData(
+                "Buffer too short for Array count".to_string(),
+            )
+            .into());
+        }
+        let count = u32::from_le_bytes(bytes[1..5].try_into().unwrap()) as usize;
+        let mut offset: usize = 5;
+
+        // Prevent DoS via memory exhaustion from malicious input
+        if count > MAX_ARRAY_ELEMENTS {
+            return Err(StorageError::CorruptedData(format!(
+                "Array count {} exceeds maximum allowed {}",
+                count, MAX_ARRAY_ELEMENTS
+            ))
+            .into());
+        }
+
+        // Prevent DoS via pre-allocation amplification:
+        // Ensure we have at least 1 byte per element in the buffer
+        // before allocating the vector.
+        if bytes.len().saturating_sub(offset) < count {
+            return Err(StorageError::CorruptedData(format!(
+                "Insufficient buffer size for Array elements: need {} bytes, have {}",
+                count,
+                bytes.len().saturating_sub(offset)
+            ))
+            .into());
+        }
+
+        let mut items = Vec::with_capacity(count);
+        for _ in 0..count {
+            if offset >= bytes.len() {
+                return Err(StorageError::CorruptedData(
+                    "Buffer exhausted while reading Array elements".to_string(),
+                )
+                .into());
+            }
+            // Recursive call with depth increment
+            let (item, consumed) =
+                PropertyValue::deserialize_recursive(&bytes[offset..], depth + 1)?;
+            items.push(item);
+            offset += consumed;
+        }
+        Ok((PropertyValue::Array(Arc::new(items)), offset))
+    }
+
+    /// Estimate the heap memory usage of this property value in bytes.
+    ///
+    /// This provides a rough estimate of heap allocations, useful for memory
+    /// accounting in tiered storage migration decisions. The estimate includes:
+    ///
+    /// - String/Bytes: actual data size (shared via Arc)
+    /// - Array: element sizes plus Vec overhead
+    /// - Vector: f32 count * 4 bytes
+    /// - SparseVector: indices + values + dimension overhead
+    ///
+    /// Note: This is an estimate. Due to Arc sharing, actual memory usage may
+    /// be lower if values are shared across versions.
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// use aletheiadb::core::PropertyValue;
+    ///
+    /// let small = PropertyValue::Int(42);
+    /// assert_eq!(small.estimated_heap_size(), 0); // No heap allocation
+    ///
+    /// let string = PropertyValue::string("hello world");
+    /// assert_eq!(string.estimated_heap_size(), 11); // String length
+    /// ```
+    pub fn estimated_heap_size(&self) -> usize {
+        // Return a large "penalty" size (10MB) on error (recursion limit exceeded).
+        // This ensures that malicious or excessively nested structures are
+        // considered "large" by cache eviction policies, rather than "small" (0),
+        // preventing them from monopolizing the cache.
+        self.estimated_heap_size_recursive(0)
+            .unwrap_or(10 * 1024 * 1024)
+    }
+
+    fn estimated_heap_size_recursive(&self, depth: usize) -> Result<usize> {
+        if depth > MAX_RECURSION_DEPTH {
+            return Err(StorageError::CorruptedData(format!(
+                "Property value recursion depth limit exceeded (max {})",
+                MAX_RECURSION_DEPTH
+            ))
+            .into());
+        }
+
+        match self {
+            PropertyValue::Null
+            | PropertyValue::Bool(_)
+            | PropertyValue::Int(_)
+            | PropertyValue::Float(_) => Ok(0),
+            PropertyValue::String(s) => Ok(s.len()),
+            PropertyValue::Bytes(b) => Ok(b.len()),
+            PropertyValue::Array(arr) => {
+                // Vec capacity overhead + recursive element sizes
+                let mut size = arr.capacity() * std::mem::size_of::<PropertyValue>();
+                for item in arr.iter() {
+                    size += item.estimated_heap_size_recursive(depth + 1)?;
+                }
+                Ok(size)
+            }
+            PropertyValue::Vector(v) => Ok(v.len() * std::mem::size_of::<f32>()),
+            PropertyValue::SparseVector(sv) => {
+                // Indices + values + SparseVec struct overhead
+                Ok(
+                    sv.nnz() * (std::mem::size_of::<u32>() + std::mem::size_of::<f32>())
+                        + std::mem::size_of::<usize>(),
+                ) // dimension field
+            }
+        }
+    }
+
+    /// Calculate the number of bytes required for serialization.
+    ///
+    /// This is used to pre-allocate buffers for serialization, avoiding reallocation
+    /// overhead during high-throughput operations (like WAL writing).
+    ///
+    /// # Size Breakdown
+    /// - Null: 1 byte
+    /// - Bool: 2 bytes
+    /// - Int: 9 bytes
+    /// - Float: 9 bytes
+    /// - String: 1 + 4 + len
+    /// - Bytes: 1 + 4 + len
+    /// - Array: 1 + 4 + sum(elements)
+    /// - Vector: 1 + 4 + (dims * 4)
+    /// - SparseVector: 1 + 4 + 4 + (nnz * 8)
+    pub fn serialized_size(&self) -> Result<usize> {
+        self.serialized_size_recursive(0)
+    }
+
+    fn serialized_size_recursive(&self, depth: usize) -> Result<usize> {
+        if depth > MAX_RECURSION_DEPTH {
+            return Err(StorageError::CorruptedData(format!(
+                "Property value recursion depth limit exceeded (max {})",
+                MAX_RECURSION_DEPTH
+            ))
+            .into());
+        }
+
+        match self {
+            PropertyValue::Null => Ok(1),
+            PropertyValue::Bool(_) => Ok(2),
+            PropertyValue::Int(_) => Ok(9),
+            PropertyValue::Float(_) => Ok(9),
+            PropertyValue::String(s) => Ok(1 + 4 + s.len()),
+            PropertyValue::Bytes(b) => Ok(1 + 4 + b.len()),
+            PropertyValue::Array(arr) => {
+                let mut elements_size = 0;
+                for v in arr.iter() {
+                    elements_size += v.serialized_size_recursive(depth + 1)?;
+                }
+                Ok(1 + 4 + elements_size)
+            }
+            PropertyValue::Vector(v) => Ok(1 + 4 + (v.len() * 4)),
+            PropertyValue::SparseVector(sv) => Ok(1 + 4 + 4 + (sv.nnz() * 8)),
+        }
+    }
+}
+
+impl Drop for PropertyValue {
+    fn drop(&mut self) {
+        // Prevent stack overflow when dropping deeply nested Arrays.
+        // We linearize the drop chain by taking ownership of children
+        // from unique Arcs and processing them iteratively.
+        if let PropertyValue::Array(arc) = self {
+            // Only process if we are the sole owner of the data (refcount == 1)
+            if let Some(vec) = Arc::get_mut(arc) {
+                // Initial work list: take all elements from the current vector
+                // std::mem::take requires Vec to implement Default, which it does (empty vec)
+                let mut stack = std::mem::take(vec);
+
+                while let Some(mut item) = stack.pop() {
+                    if let PropertyValue::Array(ref mut inner_arc) = item {
+                        // Check if we own this inner array
+                        if let Some(inner_vec) = Arc::get_mut(inner_arc) {
+                            // Move its children to the stack to be processed later
+                            // This effectively flattens the recursive structure
+                            let children = std::mem::take(inner_vec);
+                            stack.extend(children);
+                        }
+                    }
+                    // item is dropped here.
+                    // If it was a unique Array, its children were moved to stack, so it's empty.
+                    // If it was shared Array, children remain (handled by other owners).
+                    // If it was non-Array, standard drop.
+                }
+            }
+        }
+    }
+}
+
+impl PropertyValue {
+    /// Safe recursive display formatter with depth limit.
+    fn fmt_display(&self, f: &mut fmt::Formatter<'_>, depth: usize) -> fmt::Result {
+        if depth > 50 {
+            return write!(f, "...");
+        }
+
+        match self {
+            PropertyValue::Null => write!(f, "null"),
+            PropertyValue::Bool(b) => write!(f, "{}", b),
+            PropertyValue::Int(i) => write!(f, "{}", i),
+            PropertyValue::Float(fl) => write!(f, "{}", fl),
+            PropertyValue::String(s) => write!(f, "\"{}\"", s),
+            PropertyValue::Bytes(b) => write!(f, "<{} bytes>", b.len()),
+            PropertyValue::Array(a) => {
+                write!(f, "[")?;
+                for (i, v) in a.iter().enumerate() {
+                    if i > 0 {
+                        write!(f, ", ")?;
+                    }
+                    if i >= 20 {
+                        write!(f, "... ({} more)", a.len() - i)?;
+                        break;
+                    }
+                    v.fmt_display(f, depth + 1)?;
+                }
+                write!(f, "]")
+            }
+            PropertyValue::Vector(v) => write!(f, "<vector[{}]>", v.len()),
+            PropertyValue::SparseVector(sv) => {
+                write!(
+                    f,
+                    "<sparse_vector[dim={}, nnz={}]>",
+                    sv.dimension(),
+                    sv.nnz()
+                )
+            }
+        }
+    }
+
+    /// Safe recursive debug formatter with depth limit.
+    fn fmt_debug(&self, f: &mut fmt::Formatter<'_>, depth: usize) -> fmt::Result {
+        if depth > 50 {
+            return write!(f, "...");
+        }
+
+        match self {
+            PropertyValue::Null => write!(f, "Null"),
+            PropertyValue::Bool(b) => write!(f, "Bool({:?})", b),
+            PropertyValue::Int(i) => write!(f, "Int({:?})", i),
+            PropertyValue::Float(fl) => write!(f, "Float({:?})", fl),
+            PropertyValue::String(s) => write!(f, "String({:?})", s),
+            PropertyValue::Bytes(b) => write!(f, "Bytes({:?})", b),
+            PropertyValue::Array(a) => {
+                write!(f, "Array([")?;
+                for (i, v) in a.iter().enumerate() {
+                    if i > 0 {
+                        write!(f, ", ")?;
+                    }
+                    if i >= 20 {
+                        write!(f, "... ({} more)", a.len() - i)?;
+                        break;
+                    }
+                    v.fmt_debug(f, depth + 1)?;
+                }
+                write!(f, "])")
+            }
+            // Limit vector output in debug to prevent log spam
+            PropertyValue::Vector(v) => {
+                write!(f, "Vector([")?;
+                for (i, val) in v.iter().enumerate() {
+                    if i > 0 {
+                        write!(f, ", ")?;
+                    }
+                    if i >= 10 {
+                        write!(f, "... ({} more)", v.len() - i)?;
+                        break;
+                    }
+                    write!(f, "{:?}", val)?;
+                }
+                write!(f, "])")
+            }
+            PropertyValue::SparseVector(sv) => write!(f, "SparseVector({:?})", sv),
+        }
+    }
+}
+
+impl fmt::Display for PropertyValue {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.fmt_display(f, 0)
+    }
+}
+
+impl fmt::Debug for PropertyValue {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.fmt_debug(f, 0)
+    }
+}
+
+// Convenient From implementations
+impl From<bool> for PropertyValue {
+    fn from(b: bool) -> Self {
+        PropertyValue::Bool(b)
+    }
+}
+
+impl From<i64> for PropertyValue {
+    fn from(i: i64) -> Self {
+        PropertyValue::Int(i)
+    }
+}
+
+impl From<i32> for PropertyValue {
+    fn from(i: i32) -> Self {
+        PropertyValue::Int(i as i64)
+    }
+}
+
+impl From<f64> for PropertyValue {
+    fn from(f: f64) -> Self {
+        PropertyValue::Float(f)
+    }
+}
+
+impl From<String> for PropertyValue {
+    fn from(s: String) -> Self {
+        // Use Arc::from(s) directly to avoid unnecessary allocation.
+        // This leverages Rust's built-in conversion chain:
+        // String → Box<str> → Arc<str>
+        // which reuses the String's allocation instead of copying.
+        // See: https://github.com/madmax983/AletheiaDB/issues/200
+        PropertyValue::String(Arc::from(s))
+    }
+}
+
+impl From<&str> for PropertyValue {
+    fn from(s: &str) -> Self {
+        PropertyValue::String(Arc::from(s))
+    }
+}
+
+impl From<Vec<u8>> for PropertyValue {
+    fn from(b: Vec<u8>) -> Self {
+        // Use Arc::from(b) directly to avoid unnecessary allocation.
+        // This leverages Rust's built-in conversion chain:
+        // Vec<u8> → Box<[u8]> → Arc<[u8]>
+        // which reuses the Vec's allocation instead of copying.
+        // See: https://github.com/madmax983/AletheiaDB/issues/200
+        PropertyValue::Bytes(Arc::from(b))
+    }
+}
+
+impl From<&[u8]> for PropertyValue {
+    fn from(b: &[u8]) -> Self {
+        PropertyValue::Bytes(Arc::from(b))
+    }
+}
+
+impl From<Vec<PropertyValue>> for PropertyValue {
+    fn from(v: Vec<PropertyValue>) -> Self {
+        PropertyValue::Array(Arc::new(v))
+    }
+}
+
+impl From<Vec<f32>> for PropertyValue {
+    fn from(v: Vec<f32>) -> Self {
+        // Use v.into() to reuse the Vec's buffer, avoiding allocation and copy
+        PropertyValue::Vector(v.into())
+    }
+}
+
+impl From<&[f32]> for PropertyValue {
+    fn from(v: &[f32]) -> Self {
+        PropertyValue::Vector(Arc::from(v))
+    }
+}
+
+impl From<SparseVec> for PropertyValue {
+    fn from(sv: SparseVec) -> Self {
+        PropertyValue::SparseVector(Arc::new(sv))
+    }
+}
